@@ -33,153 +33,141 @@
     "ie >= 11"
   ],
   "prettier": {
-    "singleQuote": true
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "pretty-quick --staged"
-    }
+    "singleQuote": true,
+    "trailingComma": "es5"
   },
   "devDependencies": {
-    "@adlk/misty": "0.2.0",
-    "@babel/core": "7.10.4",
-    "@babel/plugin-proposal-class-properties": "7.10.4",
-    "@babel/plugin-proposal-object-rest-spread": "7.10.4",
+    "@babel/core": "7.24.5",
+    "@babel/eslint-parser": "7.24.5",
     "@babel/plugin-syntax-dynamic-import": "7.8.3",
-    "@babel/plugin-transform-runtime": "7.12.1",
-    "@babel/preset-env": "7.10.4",
-    "@babel/preset-react": "7.10.4",
-    "@babel/preset-typescript": "7.10.4",
-    "@types/classnames": "2.2.10",
-    "@types/clipboard": "2.0.1",
-    "@types/cookie": "0.4.0",
-    "@types/debug": "4.1.5",
-    "@types/electron": "1.6.10",
-    "@types/jest": "26.0.4",
-    "@types/lodash": "4.14.157",
-    "@types/react": "16.9.41",
-    "@types/react-addons-update": "0.14.20",
-    "@types/react-dom": "16.9.8",
-    "@types/react-modal": "3.10.6",
-    "@types/react-onclickoutside": "6.7.3",
-    "@types/react-redux": "7.1.9",
-    "@types/react-tabs": "2.3.2",
-    "@types/react-virtualized": "9.21.10",
-    "@types/redux": "3.6.0",
-    "@types/redux-localstorage": "1.0.8",
-    "@types/rimraf": "3.0.0",
-    "@types/simperium": "1.1.1",
-    "@types/uuid": "8.0.0",
-    "@types/valid-url": "1.0.3",
-    "@types/webdriverio": "5.0.0",
-    "@types/webpack-env": "1.15.2",
-    "@typescript-eslint/eslint-plugin": "3.6.0",
-    "@typescript-eslint/parser": "3.6.0",
-    "autoprefixer": "9.8.4",
-    "babel-eslint": "10.1.0",
-    "babel-jest": "26.1.0",
-    "babel-loader": "8.1.0",
+    "@babel/plugin-transform-class-properties": "7.24.1",
+    "@babel/plugin-transform-object-rest-spread": "7.24.5",
+    "@babel/plugin-transform-runtime": "7.24.3",
+    "@babel/preset-env": "7.24.5",
+    "@babel/preset-react": "7.24.1",
+    "@babel/preset-typescript": "7.24.1",
+    "@electron/notarize": "2.3.0",
+    "@testing-library/react": "12.1.5",
+    "@types/cookie": "0.6.0",
+    "@types/debug": "4.1.12",
+    "@types/jest": "29.5.12",
+    "@types/lodash": "4.17.1",
+    "@types/react": "17.0.19",
+    "@types/react-addons-update": "0.14.24",
+    "@types/react-dom": "17.0.11",
+    "@types/react-modal": "3.16.3",
+    "@types/react-onclickoutside": "6.7.10",
+    "@types/react-redux": "7.1.33",
+    "@types/react-virtualized": "9.21.30",
+    "@types/redux-localstorage": "1.0.12",
+    "@types/simperium": "1.1.5",
+    "@types/uuid": "9.0.8",
+    "@types/valid-url": "1.0.7",
+    "@types/webpack-env": "1.18.5",
+    "@typescript-eslint/eslint-plugin": "7.8.0",
+    "@typescript-eslint/parser": "7.8.0",
+    "autoprefixer": "10.4.19",
+    "babel-jest": "29.7.0",
+    "babel-loader": "9.1.3",
     "babel-plugin-dynamic-import-node": "2.3.3",
     "babel-plugin-lodash": "3.3.4",
     "babel-plugin-transform-react-remove-prop-types": "0.4.24",
-    "classnames": "2.2.6",
-    "concurrently": "5.2.0",
-    "css-loader": "3.6.0",
-    "debug": "4.1.1",
-    "electron": "12.0.5",
-    "electron-builder": "22.10.5",
-    "electron-notarize": "1.0.0",
-    "enzyme": "3.11.0",
-    "enzyme-adapter-react-16": "1.15.2",
-    "eslint": "7.4.0",
-    "eslint-config-prettier": "6.11.0",
-    "eslint-plugin-jest": "23.18.0",
-    "eslint-plugin-prettier": "3.1.4",
-    "eslint-plugin-react": "7.20.3",
-    "eslint-plugin-react-hooks": "4.0.5",
-    "fake-indexeddb": "3.1.0",
-    "file-loader": "6.0.0",
-    "html-webpack-plugin": "4.3.0",
-    "husky": "4.2.5",
-    "jest": "26.1.0",
-    "jest-localstorage-mock": "2.4.2",
-    "mini-css-extract-plugin": "0.9.0",
-    "node-sass": "4.14.1",
-    "postcss-loader": "3.0.0",
-    "prettier": "2.0.5",
-    "pretty-quick": "2.0.1",
-    "react-onclickoutside": "6.9.0",
-    "react-test-renderer": "16.13.1",
-    "rimraf": "3.0.2",
-    "sass-loader": "9.0.1",
-    "spectron": "11.1.0",
-    "style-loader": "1.2.1",
-    "stylelint": "13.7.1",
-    "stylelint-config-prettier": "8.0.2",
-    "stylelint-config-standard": "20.0.0",
-    "stylelint-prettier": "1.1.2",
-    "stylelint-scss": "3.18.0",
-    "typescript": "3.9.6",
-    "url-loader": "4.1.0",
-    "wait-on": "5.2.1",
-    "webpack": "4.43.0",
-    "webpack-cli": "3.3.12",
-    "webpack-dev-server": "3.11.0"
+    "classnames": "2.5.1",
+    "concurrently": "8.2.2",
+    "css-loader": "7.1.1",
+    "debug": "4.3.4",
+    "electron": "30.0.2",
+    "electron-builder": "24.13.3",
+    "eslint": "8.57.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-plugin-jest": "28.5.0",
+    "eslint-plugin-prettier": "5.1.3",
+    "eslint-plugin-react": "7.34.1",
+    "eslint-plugin-react-hooks": "4.6.2",
+    "fake-indexeddb": "5.0.2",
+    "html-webpack-plugin": "5.6.0",
+    "jest": "29.7.0",
+    "jest-environment-jsdom": "29.7.0",
+    "jest-localstorage-mock": "2.4.26",
+    "mini-css-extract-plugin": "2.9.0",
+    "postcss": "8.4.38",
+    "postcss-loader": "8.1.1",
+    "prettier": "3.2.5",
+    "pretty-quick": "4.0.0",
+    "react-onclickoutside": "6.13.0",
+    "react-test-renderer": "17.0.2",
+    "sass": "1.77.0",
+    "sass-loader": "14.2.1",
+    "style-loader": "4.0.0",
+    "stylelint": "16.5.0",
+    "stylelint-config-standard": "36.0.0",
+    "stylelint-config-standard-scss": "13.1.0",
+    "stylelint-prettier": "5.0.0",
+    "typescript": "5.4.5",
+    "wait-on": "7.2.0",
+    "webpack": "5.91.0",
+    "webpack-cli": "5.1.4",
+    "webpack-dev-server": "5.0.4"
   },
   "dependencies": {
-    "@automattic/color-studio": "2.3.0",
-    "@material-ui/core": "4.11.0",
-    "@sentry/react": "5.20.1",
-    "clipboard": "2.0.6",
-    "cookie": "0.4.1",
-    "core-js": "3.6.5",
-    "date-fns": "2.14.0",
-    "electron-fetch": "1.4.0",
-    "electron-progressbar": "1.2.0",
-    "electron-updater": "4.3.1",
+    "@automattic/color-studio": "2.6.0",
+    "@emotion/react": "11.11.4",
+    "@emotion/styled": "11.11.5",
+    "@mui/material": "5.15.16",
+    "@mui/styles": "5.15.16",
+    "@sentry/react": "7.113.0",
+    "clipboard": "2.0.11",
+    "cookie": "0.6.0",
+    "core-js": "3.37.0",
+    "date-fns": "3.6.0",
+    "electron-fetch": "1.9.1",
+    "electron-progressbar": "2.2.1",
+    "electron-updater": "6.1.8",
     "electron-window-state": "5.0.3",
-    "file-saver": "2.0.2",
-    "focus-trap-react": "^8.4.2",
-    "focus-visible": "5.1.0",
-    "gridicons": "3.3.1",
-    "highlight.js": "10.1.1",
+    "file-saver": "2.0.5",
+    "focus-trap-react": "10.2.3",
+    "focus-visible": "5.2.0",
+    "gridicons": "3.4.2",
+    "highlight.js": "11.9.0",
     "isemail": "3.2.0",
-    "js-yaml": "3.14.0",
-    "jsdom": "16.3.0",
-    "jszip": "3.5.0",
-    "lodash": "4.17.19",
+    "js-yaml": "4.1.0",
+    "jsdom": "24.0.0",
+    "jszip": "3.10.1",
+    "lodash": "4.17.21",
     "md5.js": "1.3.5",
-    "monaco-editor": "0.22.0",
-    "monaco-editor-webpack-plugin": "3.0.0",
-    "pretty-bytes": "5.3.0",
-    "promise": "8.1.0",
-    "prop-types": "15.7.2",
+    "monaco-editor": "0.48.0",
+    "monaco-editor-webpack-plugin": "7.1.0",
+    "node-polyfill-webpack-plugin": "3.0.0",
+    "pretty-bytes": "6.1.1",
+    "promise": "8.3.0",
+    "prop-types": "15.8.1",
     "randombytes": "2.1.0",
-    "react": "16.13.1",
+    "react": "17.0.2",
     "react-addons-update": "15.6.3",
-    "react-dom": "16.13.1",
-    "react-dropzone": "11.0.1",
-    "react-modal": "3.11.2",
-    "react-monaco-editor": "0.39.1",
-    "react-overlays": "3.2.0",
-    "react-redux": "7.2.0",
-    "react-sortable-hoc": "1.11.0",
-    "react-tabs": "3.1.1",
-    "react-transition-group": "4.4.1",
-    "react-virtualized": "9.21.2",
-    "redux": "4.0.5",
+    "react-dom": "17.0.2",
+    "react-dropzone": "11.7.1",
+    "react-modal": "3.16.1",
+    "react-monaco-editor": "0.50.1",
+    "react-overlays": "5.2.1",
+    "react-redux": "8.1.3",
+    "react-sortable-hoc": "2.0.0",
+    "react-tabs": "4.3.0",
+    "react-transition-group": "4.4.5",
+    "react-virtualized": "9.22.5",
+    "redux": "5.0.1",
     "redux-localstorage": "0.4.1",
-    "regenerator-runtime": "0.13.5",
-    "remove-markdown": "0.3.0",
+    "regenerator-runtime": "0.14.1",
+    "remove-markdown": "0.5.0",
     "sanitize-filename": "1.6.3",
-    "sax": "1.2.4",
-    "semver": "7.3.2",
+    "sax": "1.3.0",
+    "semver": "7.6.1",
+    "setimmediate": "1.0.5",
     "showdown": "1.9.1",
     "simperium": "1.1.4",
-    "string-replace-to-array": "1.0.3",
-    "turndown": "6.0.0",
+    "string-replace-to-array": "2.1.0",
+    "turndown": "7.1.3",
     "unorm": "1.6.0",
     "valid-url": "1.0.9",
-    "winston": "3.3.3"
+    "winston": "3.13.0"
   }
 }
