@@ -1,6 +1,5 @@
 const fs = require('fs');
 const path = require('path');
-var electron_notarize = require('electron-notarize');
 
 module.exports = async function (params) {
   // Only notarize the app on Mac OS only.
@@ -30,6 +29,7 @@ module.exports = async function (params) {
   console.log(`Notarizing ${appId} found at ${appPath}`); // eslint-disable-line no-console
 
   try {
+    const electron_notarize = require('electron-notarize');
     await electron_notarize.notarize({
       appBundleId: appId,
       appPath: appPath,
