@@ -5,6 +5,9 @@ const getConfig = require('./get-config');
 const spawnSync = require('child_process').spawnSync;
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const MonacoWebpackPlugin = require('monaco-editor-webpack-plugin');
+const NodePolyfillPlugin = require('node-polyfill-webpack-plugin');
+
+// Object.assign(global, { WebSocket: require('ws') });
 
 module.exports = () => {
   const isDevMode = process.env.NODE_ENV === 'development';
@@ -14,22 +17,21 @@ module.exports = () => {
     context: __dirname + '/lib',
     mode: isDevMode ? 'development' : 'production',
     devtool:
-      process.env.SOURCEMAP || (isDevMode && 'cheap-module-eval-source-map'),
-    devServer: { inline: true },
+      process.env.SOURCEMAP || (isDevMode && 'eval-cheap-module-source-map'),
     entry: ['./boot'],
     output: {
-      path: __dirname + '/dist',
-      filename: 'app.[hash].js',
+      filename: 'app.[fullhash].js',
       chunkFilename: '[name].[chunkhash].js',
       ...(config.is_app_engine && {
         publicPath: config.web_app_url + '/',
       }),
     },
+    // target: 'node',
     module: {
       rules: [
         {
           test: /\.[jt]sx?$/,
-          exclude: /node_modules\/core-js/,
+          exclude: /node_modules/,
           use: [
             {
               loader: 'babel-loader',
@@ -49,7 +51,9 @@ module.exports = () => {
             {
               loader: 'postcss-loader',
               options: {
-                plugins: [autoprefixer()],
+                postcssOptions: {
+                  plugins: [autoprefixer()],
+                },
                 sourceMap: isDevMode,
               },
             },
@@ -66,15 +70,21 @@ module.exports = () => {
         },
         {
           test: /\.ttf$/,
-          use: ['file-loader'],
+          type: 'asset/resource',
         },
       ],
     },
     resolve: {
+      // fallback: {
+      //   setImmediate: require.resolve('setimmediate/'),
+      // },
       extensions: ['.js', '.jsx', '.json', '.scss', '.css', '.ts', '.tsx'],
       modules: ['node_modules'],
     },
     plugins: [
+      new NodePolyfillPlugin({
+        includeAliases: ['path', 'process', 'stream', 'util'],
+      }),
       new HtmlWebpackPlugin({
         'build-platform': process.platform,
         'build-reference': spawnSync('git', ['describe', '--always', '--dirty'])
@@ -86,8 +96,8 @@ module.exports = () => {
         title: 'Simplenote',
       }),
       new MiniCssExtractPlugin({
-        filename: isDevMode ? '[name].css' : '[name].[hash].css',
-        chunkFilename: isDevMode ? '[id].css' : '[id].[hash].css',
+        filename: isDevMode ? '[name].css' : '[name].[fullhash].css',
+        chunkFilename: isDevMode ? '[id].css' : '[id].[fullhash].css',
       }),
       new MonacoWebpackPlugin({
         languages: [],
@@ -118,7 +128,16 @@ module.exports = () => {
         __TEST__: JSON.stringify(process.env.NODE_ENV === 'test'),
         config: JSON.stringify(config),
       }),
-      new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
+      new webpack.IgnorePlugin({
+        resourceRegExp: /^\.\/locale$/,
+        contextRegExp: /moment$/,
+      }),
+      // new webpack.ProvidePlugin({
+      //   setImmediate: require.resolve('setimmediate/'),
+      // }),
+      // new webpack.ProvidePlugin({
+      //   isemail: require.resolve('isemail/'),
+      // }),
     ],
   };
 };
